# docs/swagger.yaml
basePath: /api
definitions:
  dto.AuthResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      password:
        example: P@ssw0rd!
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      fullName:
        example: Alice Doe
        type: string
      password:
        example: P@ssw0rd!
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Quasar frontend'in kullandığı Go backend API dokümantasyonu.
  title: My API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Email + password ile giriş
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Yeni kullanıcı kaydı (email, password, fullName)
      parameters:
      - description: Register payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Register user
      tags:
      - Auth
  /health:
    get:
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Health check
      tags:
      - Health
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"